pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentlens"
version = "0.1.0"
description = "Pre-Installation AI Agent Tool Risk Analyzer"
dependencies = [
    "click>=8.1.7",
    "openai>=1.0.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1"
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "AgentLens contributors" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Security",
]

[project.optional-dependencies]
injection-prefilter = [
    "transformers>=4.36.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/ellacarmon/AgentLens"
Repository = "https://github.com/ellacarmon/AgentLens"

[project.scripts]
agentlens = "agentlens.cli:main"

[tool.setuptools.packages.find]
include = ["agentlens*"]
